[REFACT]: reorganizing functions (#110)
* refactor: removing the accessKey interface from lambda's interfaces

* feat: add AccessKey and AccessKeyParam on secretsManager interfaces

* feat: reorganizing secretsManager folder and functions

* refactor: improving invoke function

* feat: update version

* fix: ignoring SecretManagger functions

* Update src/lambda/lambdaService.ts
