+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "adapcon-utils-js",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "description": "Utils library for Javascript",
   "keywords": [],
   "author": {
