+++ src/lambda/lambdaService.ts
@@ -1,23 +1,16 @@
 import AWS from 'aws-sdk'
 
 import { formattedResponse } from './formatters'
-import { SecretManager } from './../secretsManager/index'
-import { lambdaParameters, AccessKey } from '.'
+import { SecretManager } from './../secretsManager'
+import { lambdaParameters } from '.'
 
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
 export class LambdaService {
   static async invoke (lambdaParameters: lambdaParameters): Promise<object> {
     try {
-      const accessKey: AccessKey = lambdaParameters.serviceSecretArn
-        ? await SecretManager.getValue({
-          region: lambdaParameters.region,
-          secretId: lambdaParameters.serviceSecretArn
-        })
-        : {}
-
       return executeInvoke({
         ...lambdaParameters,
-        ...accessKey
+        ...await SecretManager.getAccessKey(lambdaParameters)
       })
     } catch (error) {
       console.log('Error LambdaService invoke', error)
