+++ src/secretsManager/interfaces.ts
@@ -1,0 +1,10 @@
+export interface AccessKey {
+  accessKeyId: string
+  secretAccessKey: string
+  region?: string
+}
+
+export interface AccessKeyParam {
+  region?: string
+  serviceSecretArn?: string
+}
